# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Goa Trust Realty Backend API",
    "description": "API documentation for the Goa Trust Realty property and listing modules. Includes full CRUD capabilities and advanced filtering for property listings.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/properties": {
      "get": {
        "summary": "List all properties",
        "description": "Retrieve property listings with optional search and filters.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search keyword matching title or location",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "in": "query",
            "description": "Filter by location name (e.g. Assagao)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "listingType",
            "in": "query",
            "description": "Filter by listing contract type",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["sale", "rent"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of listings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Property"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a property listing",
        "description": "Scaffolds and registers a new property in the listings catalog.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Property listing created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "$ref": "#/components/schemas/Property"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation errors from payload sanitization"
          },
          "409": {
            "description": "Duplicate slug conflict error"
          }
        }
      }
    },
    "/api/properties/{id}": {
      "get": {
        "summary": "Get property details",
        "description": "Retrieve property information by either its UUID or unique URL slug.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID or URL slug of the property",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property details successfully found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "$ref": "#/components/schemas/Property"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Property not found"
          }
        }
      },
      "put": {
        "summary": "Update property listing",
        "description": "Perform full or partial updates to an existing property listing.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID or URL slug of the property",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "$ref": "#/components/schemas/Property"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Property not found"
          }
        }
      },
      "delete": {
        "summary": "Remove property listing",
        "description": "Permanently deletes a property from the system catalog.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID or URL slug of the property",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "$ref": "#/components/schemas/Property"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Property not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Property": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "location": { "type": "string" },
          "price": { "type": "string" },
          "image": { "type": "string" },
          "bedrooms": { "type": "integer" },
          "bathrooms": { "type": "integer" },
          "area": { "type": "string" },
          "listingType": { "type": "string", "enum": ["sale", "rent"] },
          "verified": { "type": "boolean" },
          "slug": { "type": "string" }
        }
      },
      "CreatePropertyInput": {
        "type": "object",
        "required": ["title", "location", "price", "image", "bedrooms", "bathrooms", "area", "listingType", "slug"],
        "properties": {
          "title": { "type": "string", "minLength": 3 },
          "location": { "type": "string", "minLength": 3 },
          "price": { "type": "string" },
          "image": { "type": "string" },
          "bedrooms": { "type": "integer", "minimum": 0 },
          "bathrooms": { "type": "integer", "minimum": 0 },
          "area": { "type": "string" },
          "listingType": { "type": "string", "enum": ["sale", "rent"] },
          "verified": { "type": "boolean", "default": false },
          "slug": { "type": "string", "pattern": "^[a-z0-9-]+$" }
        }
      },
      "UpdatePropertyInput": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 3 },
          "location": { "type": "string", "minLength": 3 },
          "price": { "type": "string" },
          "image": { "type": "string" },
          "bedrooms": { "type": "integer", "minimum": 0 },
          "bathrooms": { "type": "integer", "minimum": 0 },
          "area": { "type": "string" },
          "listingType": { "type": "string", "enum": ["sale", "rent"] },
          "verified": { "type": "boolean" },
          "slug": { "type": "string", "pattern": "^[a-z0-9-]+$" }
        }
      }
    }
  }
}